fix: refactor comment to SDL
