+++ packages/server/graphql/public/types/UpdateCommentContentSuccess.ts
@@ -1,0 +1,13 @@
+import {UpdateCommentContentSuccessResolvers} from '../resolverTypes'
+
+export type UpdateCommentContentSuccessSource = {
+  commentId: string
+}
+
+const UpdateCommentContentSuccess: UpdateCommentContentSuccessResolvers = {
+  comment: async ({commentId}, _args, {dataLoader}) => {
+    return dataLoader.get('comments').load(commentId)
+  }
+}
+
+export default UpdateCommentContentSuccess
