+++ packages/server/graphql/rootMutation.ts
@@ -19,7 +19,6 @@ import createPoll from './mutations/createPoll'
 import createReflection from './mutations/createReflection'
 import createTask from './mutations/createTask'
 import createTaskIntegration from './mutations/createTaskIntegration'
-import deleteComment from './mutations/deleteComment'
 import deleteTask from './mutations/deleteTask'
 import deleteUser from './mutations/deleteUser'
 import denyPushInvitation from './mutations/denyPushInvitation'
@@ -91,7 +90,6 @@ import startDraggingReflection from './mutations/startDraggingReflection'
 import startSprintPoker from './mutations/startSprintPoker'
 import toggleTeamDrawer from './mutations/toggleTeamDrawer'
 import updateAzureDevOpsDimensionField from './mutations/updateAzureDevOpsDimensionField'
-import updateCommentContent from './mutations/updateCommentContent'
 import updateDragLocation from './mutations/updateDragLocation'
 import updateGitHubDimensionField from './mutations/updateGitHubDimensionField'
 import updateNewCheckInQuestion from './mutations/updateNewCheckInQuestion'
@@ -130,7 +128,6 @@ export default new GraphQLObjectType<any, GQLContext>({
       createOAuth1AuthorizeUrl,
       createReflection,
       createTask,
-      deleteComment,
       deleteTask,
       deleteUser,
       denyPushInvitation,
@@ -188,7 +185,6 @@ export default new GraphQLObjectType<any, GQLContext>({
       startDraggingReflection,
       startSprintPoker,
       setTaskHighlight,
-      updateCommentContent,
       oldUpdateCreditCard,
       updatePokerTemplateDimensionScale,
       updatePokerTemplateScaleValue,
